+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+- 1.2.0
+  - Updated model descriptions and failure messages.
+
+- 1.1.0
+  - Added model association matchers.
+  - Added model attribute matchers.
+
 - 1.0.1
   - Fixed issue where matchers would error unless explicitly passed the controller object.